refactor: reestructure /deputado-federal components directory
Margin from footer was fixed also
