--- app/components/Footer.js
@@ -1,9 +1,0 @@
-export default function Footer() {
-  return (
-    <div className="flex items-center justify-center w-full px-2 py-2">
-      <span className="text-small text-default-400">
-        Criado com amor por brasileiros.
-      </span>
-    </div>
-  );
-}
