+++ app/deputado-federal/_components/index.js
@@ -1,0 +1,3 @@
+import { DeputiesTable } from "./DeputiesTable";
+
+export { DeputiesTable };
